+++ server/src/server.ts
@@ -1,5 +1,6 @@
 import Fastify, { FastifyServerOptions, FastifyLoggerInstance } from 'fastify';
 import { Server } from 'http';
+import { createAnalyticsHistoricalEndpoint } from './api/modules/analytics/historical';
 import { createAnalyticsReferrersEndpoint } from './api/modules/analytics/referrers';
 
 export const buildFastifyServer = (
@@ -17,6 +18,7 @@ export const buildFastifyServer = (
    * Analytics routes
    */
   createAnalyticsReferrersEndpoint(fastify);
+  createAnalyticsHistoricalEndpoint(fastify);
 
   return fastify;
 };
